move cluster fixture to own module
